django_app/requirements.txt
django==4.2.13
python-dotenv==1.0.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
httpx==0.27.0
gunicorn==22.0.0
whitenoise==6.7.0
